Allow sharing and sending bitmaps across threads.
This involves careful tracking of which bitmaps are targeted where. In general,
you cannot target a bitmap in multiple threads, or destroy a bitmap targeted in
another thread. Same thing applies to displays.

This also implies that the target bitmap can be null, so various drawing
commands can now panic.